Moved 'CalendarDay' out of SimpleMonthAdapter
