+++ datetimepicker-library/src/com/fourmob/datetimepicker/date/SimpleMonthAdapter.java
@@ -1,8 +1,6 @@
 package com.fourmob.datetimepicker.date;
 
 import android.content.Context;
-import android.os.Parcel;
-import android.os.Parcelable;
 import android.util.Log;
 import android.view.View;
 import android.view.ViewGroup;
@@ -10,10 +8,6 @@ import android.view.ViewGroup.LayoutParams;
 import android.widget.AbsListView;
 import android.widget.BaseAdapter;
 
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Calendar;
-import java.util.Date;
 import java.util.HashMap;
 
 public class SimpleMonthAdapter extends BaseAdapter implements SimpleMonthView.OnDayClickListener {
@@ -140,112 +134,4 @@ public class SimpleMonthAdapter extends BaseAdapter implements SimpleMonthView.O
         notifyDataSetChanged();
     }
 
-    public static class CalendarDay implements Parcelable {
-        private Calendar calendar;
-
-        int day;
-        int month;
-        int year;
-
-        public CalendarDay() {
-            setTime(System.currentTimeMillis());
-        }
-
-        public CalendarDay(int year, int month, int day) {
-            setDay(year, month, day);
-        }
-
-        public CalendarDay(long timeInMillis) {
-            setTime(timeInMillis);
-        }
-
-        public CalendarDay(Calendar calendar) {
-            year = calendar.get(Calendar.YEAR);
-            month = calendar.get(Calendar.MONTH);
-            day = calendar.get(Calendar.DAY_OF_MONTH);
-        }
-
-        private CalendarDay(Parcel in) {
-            year = in.readInt();
-            month = in.readInt();
-            day = in.readInt();
-        }
-
-        public static final Creator CREATOR = new Creator() {
-            public CalendarDay createFromParcel(Parcel in) {
-                return new CalendarDay(in);
-            }
-
-            public CalendarDay[] newArray(int size) {
-                return new CalendarDay[size];
-            }
-        };
-
-        @Override
-        public int describeContents() {
-            return 0;
-        }
-
-        @Override
-        public void writeToParcel(Parcel dest, int flags) {
-            dest.writeInt(year);
-            dest.writeInt(month);
-            dest.writeInt(day);
-        }
-
-        private void setTime(long timeInMillis) {
-            if (calendar == null) {
-                calendar = Calendar.getInstance();
-            }
-            calendar.setTimeInMillis(timeInMillis);
-            month = this.calendar.get(Calendar.MONTH);
-            year = this.calendar.get(Calendar.YEAR);
-            day = this.calendar.get(Calendar.DAY_OF_MONTH);
-        }
-
-        public void set(CalendarDay calendarDay) {
-            year = calendarDay.year;
-            month = calendarDay.month;
-            day = calendarDay.day;
-        }
-
-        public void setDay(int year, int month, int day) {
-            this.year = year;
-            this.month = month;
-            this.day = day;
-        }
-
-        public boolean equals(CalendarDay o) {
-            if(o.year == year && o.day == day && o.month == month){
-                return true;
-            }
-            else return false;
-        }
-
-        /**
-         * returns true if day is after day
-         *
-         * @param day
-         * @return
-         */
-        public boolean isAfter(CalendarDay day) {
-            return convertToDate(this).after(convertToDate(day));
-        }
-
-        public boolean isBefore(CalendarDay day) {
-            return convertToDate(this).before(convertToDate(day));
-        }
-
-        public Date convertToDate(CalendarDay day) {
-            java.util.Date utilDate = null;
-
-            try {
-                SimpleDateFormat formatter = new SimpleDateFormat("yyyy/MM/dd");
-                utilDate = formatter.parse(day.year + "/" + day.month + "/" + day.day);
-            } catch (ParseException e) {
-                Log.e("SimpleMonthAdapter", "error while converting", e);
-            }
-            return utilDate;
-        }
-    }
 }
